PR updates: docs, file structure, cluster factory
